Move chain plots and convergence plots to analysis results class so they can be easily plotted from analysis results written to disk. Fix number of samples used in convergence plots (bootstrap histogram)
